+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 # v1.2 - x/06/2017
 - Add error on invalid currency
+- Fix an error on tip to lowercase user registered
 
 # v1.1 - 16/06/2017
 - Add support of new syntax for tip command #57